pyproject.toml
[tool.poetry]
name = "realway"
version = "0.1.0"
description = ""
authors = ["XuQiang <xq_work@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.1"
rich = "^12.5.1"
tomli = "^2.0.1"
pymongo = "^4.2.0"
Flask = "^2.2.2"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pypinyin = "^0.46.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
